+++ backend/init.php
@@ -37,11 +37,6 @@ class BackendInit
 		// register the autoloader
 		spl_autoload_register(array('BackendInit', 'autoLoader'));
 
-		// set some ini-options
-		ini_set('pcre.backtrack_limit', 999999999);
-		ini_set('pcre.recursion_limit', 999999999);
-		ini_set('memory_limit', '64M');
-
 		// set a default timezone if no one was set by PHP.ini
 		if(ini_get('date.timezone') == '') date_default_timezone_set('Europe/Brussels');