+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Improvements:
 * Core: Upgraded CKFinder to 2.3.0
 * Formbuilder: added API-methods.
 * FAQ: Setting for "one category" added, so user only has 1 category in the website + the category title is hidden in the frontend (for smaller websites).
+* Core: removed ini_set's. Let the server handle this.
 
 Bugfixes: